requirements.txt
anyio==3.6.2
asyncio-mqtt==0.13.0
click==8.1.3
dht11==0.1.0
dotenv==0.0.5
fastapi==0.86.0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
Jinja2==3.1.2
MarkupSafe==2.1.1
paho-mqtt==1.6.1
peewee==3.15.3
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.21.0
RPi.GPIO==0.7.1
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
zipp==3.10.0
